requirements.txt
Appium-Python-Client
pytest
behave~=1.2.6
selenium
webdriver-manager
pyyaml
PyHamcrest
selenium-page-factory